fix: share the DummyCircuit betw tests & prover
