+++ fastcrypto-zkp/benches/proving.rs
@@ -5,50 +5,12 @@ use ark_crypto_primitives::SNARK;
 use ark_ec::PairingEngine;
 use ark_ff::{PrimeField, UniformRand};
 use ark_groth16::Groth16;
-use ark_relations::{
-    lc,
-    r1cs::{ConstraintSynthesizer, ConstraintSystemRef, SynthesisError},
-};
 use criterion::{
     criterion_group, criterion_main, measurement::Measurement, BenchmarkGroup, BenchmarkId,
     Criterion, SamplingMode,
 };
+use fastcrypto_proving::dummy_circuit::DummyCircuit;
 use std::ops::Mul;
-
-#[derive(Copy, Clone)]
-struct DummyCircuit<F: PrimeField> {
-    pub a: Option<F>,
-    pub b: Option<F>,
-    pub num_variables: usize,
-    pub num_constraints: usize,
-}
-
-impl<F: PrimeField> ConstraintSynthesizer<F> for DummyCircuit<F> {
-    // We'll be proving a relationship involving the product c of a & b
-    fn generate_constraints(self, cs: ConstraintSystemRef<F>) -> Result<(), SynthesisError> {
-        let a = cs.new_witness_variable(|| self.a.ok_or(SynthesisError::AssignmentMissing))?;
-        let b = cs.new_witness_variable(|| self.b.ok_or(SynthesisError::AssignmentMissing))?;
-        let c = cs.new_input_variable(|| {
-            let a = self.a.ok_or(SynthesisError::AssignmentMissing)?;
-            let b = self.b.ok_or(SynthesisError::AssignmentMissing)?;
-
-            Ok(a * b)
-        })?;
-
-        // a, b, c are above, let's define the rest.
-        for _ in 0..(self.num_variables - 3) {
-            let _ = cs.new_witness_variable(|| self.a.ok_or(SynthesisError::AssignmentMissing))?;
-        }
-
-        for _ in 0..self.num_constraints - 1 {
-            cs.enforce_constraint(lc!() + a, lc!() + b, lc!() + c)?;
-        }
-
-        cs.enforce_constraint(lc!(), lc!(), lc!())?;
-
-        Ok(())
-    }
-}
 
 fn bench_prove<F: PrimeField, E: PairingEngine<Fr = F>, M: Measurement>(
     grp: &mut BenchmarkGroup<M>,
