+++ fastcrypto-zkp/src/lib.rs
@@ -1,4 +1,11 @@
 // Copyright (c) 2022, Mysten Labs, Inc.
 // SPDX-License-Identifier: Apache-2.0
+
+// Conversions between arkworks <-> blst
 pub mod conversions;
+
+// dummy circuit used for tests and benchmarks
+pub mod dummy_circuit;
+
+// Groth16 SNARK verifier
 pub mod verifier;
