+++ fastcrypto-zkp/src/unit_tests/verifier_tests.rs
@@ -15,10 +15,6 @@ use ark_crypto_primitives::SNARK;
 use ark_ec::{AffineCurve, PairingEngine, ProjectiveCurve};
 use ark_ff::{One, PrimeField, UniformRand};
 use ark_groth16::{Groth16, PreparedVerifyingKey};
-use ark_relations::{
-    lc,
-    r1cs::{ConstraintSynthesizer, ConstraintSystemRef, SynthesisError},
-};
 
 use crate::{
     conversions::{
@@ -26,49 +22,14 @@ use crate::{
         bls_g2_affine_to_blst_g2_affine, blst_fp12_to_bls_fq12,
         tests::{arb_bls_fr, arb_bls_g1_affine, arb_blst_g1_affine, arb_blst_g2_affine},
     },
+    dummy_circuit::DummyCircuit,
     verifier::{
         g1_linear_combination, multipairing_with_processed_vk, process_vk_special,
         verify_with_processed_vk, Proof, VerifyingKey,
     },
 };
 
-// This duplicates the proving DummyCircuit, which is a bit hard to share.
-#[derive(Copy, Clone)]
-struct DummyCircuit<F: PrimeField> {
-    pub a: Option<F>,
-    pub b: Option<F>,
-    pub num_variables: usize,
-    pub num_constraints: usize,
-}
-
-impl<F: PrimeField> ConstraintSynthesizer<F> for DummyCircuit<F> {
-    // We'll be proving a relationship involving the product c of a & b
-    fn generate_constraints(self, cs: ConstraintSystemRef<F>) -> Result<(), SynthesisError> {
-        let a = cs.new_witness_variable(|| self.a.ok_or(SynthesisError::AssignmentMissing))?;
-        let b = cs.new_witness_variable(|| self.b.ok_or(SynthesisError::AssignmentMissing))?;
-        let c = cs.new_input_variable(|| {
-            let a = self.a.ok_or(SynthesisError::AssignmentMissing)?;
-            let b = self.b.ok_or(SynthesisError::AssignmentMissing)?;
-
-            Ok(a * b)
-        })?;
-
-        // a, b, c are above, let's define the rest.
-        for _ in 0..(self.num_variables - 3) {
-            let _ = cs.new_witness_variable(|| self.a.ok_or(SynthesisError::AssignmentMissing))?;
-        }
-
-        for _ in 0..self.num_constraints - 1 {
-            cs.enforce_constraint(lc!() + a, lc!() + b, lc!() + c)?;
-        }
-
-        cs.enforce_constraint(lc!(), lc!(), lc!())?;
-
-        Ok(())
-    }
-}
-
-// This emulates the process_vk function of the arkworks verifier, but using blst to compute the term 
+// This emulates the process_vk function of the arkworks verifier, but using blst to compute the term
 // alpha_g1_beta_g2. See [`test_prepare_vk`].
 fn ark_process_vk(vk: &VerifyingKey<Bls12_381>) -> PreparedVerifyingKey<Bls12_381> {
     let g1_alpha = bls_g1_affine_to_blst_g1_affine(&vk.alpha_g1);
@@ -91,7 +52,7 @@ fn ark_process_vk(vk: &VerifyingKey<Bls12_381>) -> PreparedVerifyingKey<Bls12_38
 }
 
 // This commputes the result of the multi-pairing involved in the Groth16 verification, using arkworks.
-// See [`test_multipairing_with_processed_vk`] 
+// See [`test_multipairing_with_processed_vk`]
 pub fn ark_multipairing_with_prepared_vk(
     pvk: &PreparedVerifyingKey<Bls12_381>,
     proof: &Proof<Bls12_381>,
